Move to heatmap routes into their own class
