+++ app_backend/src/metta/app_backend/server.py
@@ -9,7 +9,7 @@ from fastapi.middleware.cors import CORSMiddleware
 
 from metta.app_backend.auth import user_from_header_or_token
 from metta.app_backend.metta_repo import MettaRepo
-from metta.app_backend.routes import dashboard_routes, sql_routes, stats_routes, token_routes
+from metta.app_backend.routes import dashboard_routes, heatmap_routes, sql_routes, stats_routes, token_routes
 
 _logging_configured = False
 
@@ -85,11 +85,13 @@ def create_app(stats_repo: MettaRepo) -> fastapi.FastAPI:
     sql_router = sql_routes.create_sql_router(stats_repo)
     stats_router = stats_routes.create_stats_router(stats_repo)
     token_router = token_routes.create_token_router(stats_repo)
+    heatmap_router = heatmap_routes.create_heatmap_router(stats_repo)
 
     app.include_router(dashboard_router)
     app.include_router(sql_router)
     app.include_router(stats_router)
     app.include_router(token_router)
+    app.include_router(heatmap_router)
 
     @app.get("/whoami")
     async def whoami(request: fastapi.Request):  # type: ignore
